Move code to helper
